% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptools_financialfunctions.r
\name{amort_cd}
\alias{amort_cd}
\title{Constant dollar amortization method}
\usage{
amort_cd(p, i, m, end)
}
\arguments{
\item{p}{principal}

\item{i}{interest rate}

\item{m}{number of periods}

\item{end}{if TRUE, payment at the end of period}
}
\value{
The amortization payment
}
\description{
\code{amort_cd} initial payment for a graduated annuity growing at rate g
}
\details{
Returns POSITIVE payment for positive principal
}
\examples{
amort_cd(100, 0.08, 10, FALSE)
}
\keyword{amort_cd}
