% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptools_financialfunctions.r
\name{amort_cp}
\alias{amort_cp}
\title{Constant percent amortization method}
\usage{
amort_cp(p, i, m, g, end)
}
\arguments{
\item{p}{principal}

\item{i}{interest rate}

\item{m}{number of periods}

\item{g}{growth rate}

\item{end}{if TRUE, payment at the end of period}
}
\value{
The amortization payment
}
\description{
\code{amort_cp} stream of payments for a graduated annuity growing at rate g
}
\details{
Returns POSITIVE payment for positive principal
}
\examples{
amort_cp(100, 0.08, 10, .03, FALSE)
}
\keyword{amort_cp}
