% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptools_financialfunctions.r
\name{amort_LG}
\alias{amort_LG}
\title{Function for choosing amortization methods}
\usage{
amort_LG(p, i, m, g, end, method)
}
\arguments{
\item{p}{principal}

\item{i}{interest rate}

\item{m}{number of periods}

\item{g}{growth rate}

\item{end}{if TRUE, payment at the end of period}

\item{method}{one of c("cd", "cp", "sl")}
}
\value{
The stream of payments
}
\description{
\code{amort_LG} stream of payments for straight-line payments
}
\details{
Returns stream of payments
}
\examples{
amort_LG(100, 0.08, 10, .03, method="cd")
amort_LG(100, 0.08, 10, .03, method="cp")
amort_LG(100, 0.08, 10, .03, method="sl")
}
\keyword{amort_LG}
