% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptools_financialfunctions.r
\name{amort_sl}
\alias{amort_sl}
\title{Constant percent amortization method}
\usage{
amort_sl(p, i, m, end)
}
\arguments{
\item{p}{principal}

\item{i}{interest rate}

\item{m}{number of periods}

\item{end}{if TRUE, payment at the end of period}
}
\value{
The amortization payment
}
\description{
\code{amort_sl} stream of payments for straight-line payments
}
\details{
Returns POSITIVE payment for positive principal
}
\examples{
amort_sl(100, 0.08, 10, FALSE)

amort_cd(100, 0.08, 10, FALSE)
amort_cp(100, 0.08, 10, 0.05, FALSE)
amort_sl(100, 0.08, 10, FALSE)
}
\keyword{amort_sl}
