% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplotStack}
\alias{barplotStack}
\title{Stack Barplot by Group}
\usage{
barplotStack(
  data,
  groupinfo = NULL,
  horiz = F,
  args.group = NULL,
  args.legend = NULL,
  ...
)
}
\arguments{
\item{data}{A matrix with the sample ids as colnames and the cell types as rownames.}

\item{groupinfo}{A vector/factor or list specifies which group the samples belong to.}

\item{horiz}{a logical value. If FALSE, the bars are drawn vertically with the first
bar to the left. If TRUE, the bars are drawn horizontally with the first at the bottom}

\item{args.group}{list of additional arguments to pass to text(), which adds group labels
on plot. names of the list are used as argument names.}

\item{args.legend}{list of additional arguments to pass to legend().
names of the list are used as argument names.}

\item{...}{arguments to be passed to/from other methods,
such as \code{barplot()} or \code{par()}.}
}
\value{
A numeric vector, say mp, giving the coordinates of all the bar midpoints drawn,
useful for adding to the graph.
}
\description{
Creates a bar plot with vertical or horizontal stacked bars clustered according to
group information.
}
\examples{
cellprop <- matrix(runif(15 * 6), ncol = 6)
rownames(cellprop) <- paste0("S", seq_len(nrow(cellprop)))
colnames(cellprop) <- c("B", "CD4", "CD8", "Mono/Macro", "NK", "Neutro")
cellprop <- cellprop/rowSums(cellprop)
cellprop <- t(cellprop)
print(cellprop)
groupinfo <- sample(c("Group1", "Group2", "Group3"), ncol(cellprop), replace = TRUE)
barplotStack(cellprop, groupinfo = groupinfo, horiz = TRUE)
barplotStack(cellprop, groupinfo = groupinfo)
}
\author{
Yuhao Xie
}
