% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_init.R
\name{connect_init}
\alias{connect_init}
\title{connect_init}
\usage{
connect_init(id, pw, host, port)
}
\arguments{
\item{id}{, identifiaction credential for connexion}

\item{pw}{, password credential for connexion}

\item{host}{, an IP adress}

\item{port}{, the Impala (or Hive) port to connect to}
}
\value{
a connection to impala, the driver (depending on OS) is automatically detected.
}
\description{
connect_init
}
