% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrPairs.R
\name{corrPairs}
\alias{corrPairs}
\title{A pairs visualization of a correlation matrix.}
\usage{
corrPairs(
  m,
  cor.method = c("pearson", "kendall", "spearman"),
  colors = colorRampPalette(c("#CCCCCC", "#EF3B3C"))(50),
  cor.use = c("all.obs", "complete.obs", "pairwise.complete.obs", "everything",
    "na.or.complete"),
  max.col = NULL,
  min.col = NULL,
  axt = FALSE,
  cex.coeff = NULL,
  cex.point = 1,
  cex.labels = 1,
  cex.main = 1,
  line.main = 3,
  ...
)
}
\arguments{
\item{m}{a data matrix, each sample is represent by a column.}

\item{cor.method}{a character string indicating which correlation coefficient (or covariance) is to be computed.
One of "pearson" (default), "kendall", or "spearman": can be abbreviated.}

\item{colors}{colors that mapped to the coffeicient of correlation}

\item{cor.use}{an optional character string giving a method for computing covariances in the presence of missing values.
This must be (an abbreviation of) one of the strings "everything", "all.obs", "complete.obs", "na.or.complete", or
"pairwise.complete.obs".}

\item{max.col}{a numeric between 0-1, specifiy the maximal correlation coffeicient, used to scale the color.}

\item{min.col}{a numeric between 0-1, specifiy the minimal correlation coffeicient, used to scale the color.}

\item{axt}{logical. Should the xaxis and yaxis be drawn}

\item{cex.coeff}{graphics parameters for the correlation coefficient plot panel, defines the font size}

\item{cex.point}{graphics parameters for the scatter plot panel, defines the point size}

\item{cex.labels}{graphics parameters for the text panel, defines the font size}

\item{cex.main}{graphics parameters for the title, defines the font size}

\item{line.main}{if \code{main} is specified, \code{line.main} gives the line argument to \code{\link[=mtext]{mtext()}} which draws the title.
You may want to specify oma when changing line.main.}

\item{...}{other parameters pass to \code{pairs()} function}
}
\description{
A graphical display of a correlation matrix, based on \code{pairs()} function
}
\examples{
x <- rnorm(n = 100, mean = 10, sd = 2)
df <- data.frame(x = x + rnorm(100))
for (i in 1:9) {
  df <- cbind(df, x + rnorm(100))
}
colnames(df) <- paste0(LETTERS[1:10], 1:10)

corrPairs(df, axt = FALSE)
}
