% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.Interactions.R
\name{count.Interactions}
\alias{count.Interactions}
\title{count the pairs of consecutive variables in a ranger Random Forest}
\usage{
count.Interactions(object)
}
\arguments{
\item{object}{a fitted RF model of class ranger

@return  countMatrix  the counts
@return  InteractionMatrix}
}
\description{
count the pairs of consecutive variables in a ranger RF
}
\examples{
model.rf <- ranger(Species ~ ., data = iris,num.trees=500,seed =1)
countMatrix <- count.Interactions(model.rf)

}
