% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ptools_utility_functions.r
\name{cton}
\alias{cton}
\title{Character to numeric}
\usage{
cton(cvec)
}
\arguments{
\item{cvec}{character vector that we want to convert to numeric}
}
\value{
numeric vector
}
\description{
\code{cton} returns a numeric vector, converted from character
}
\details{
Eliminates "," "$" "%"  Other chars will become NA
}
\examples{
cvec <- c("12345678", "123.45678", "$123,456.78", "A123,456.78")
cton(cvec)
}
\keyword{cton}

