% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityByGroup.R
\name{densityByGroup.classic}
\alias{densityByGroup.classic}
\title{classic density plot}
\usage{
densityByGroup.classic(
  x,
  col = NULL,
  alpha = 0.5,
  args.density = NULL,
  args.legend = NULL,
  ...
)
}
\arguments{
\item{x}{a numeric vector (representing one group), or a list containing
several numeric vectors, with each representing a group.
If it is a list, its names determines the group names}

\item{col}{colors filled under density curve.}

\item{alpha}{a number modifying the opacity alpha; typically in \code{[0,1]}.}

\item{args.density}{list of additional arguments to pass to \code{density()}.
names of the list are used as argument names.}

\item{args.legend}{list of additional arguments to pass to \code{legend()}.
names of the list are used as argument names.}

\item{...}{arguments to be passed to \code{plot()}.}
}
\description{
Plot several density plot in one figure.
}
\examples{
xls <- list(
    male = rnorm(100, 170, 5),
    female = rnorm(110, 160, 5)
)
print(str(xls))
#densityByGroup.classic(xls)

}
