% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_2_impala.R
\name{df_2_impala}
\alias{df_2_impala}
\title{df_2_impala}
\usage{
df_2_impala(hdfsUri, dirUri, db_name, tab_name, id, pw, host, port,
  delim = ",")
}
\arguments{
\item{dirUri}{, directory of interest}

\item{db_name}{, database name}

\item{tab_name}{, name of the table to be created}

\item{id}{, identification credential}

\item{pw}{, password credential}

\item{host}{, url of impala/hive service}

\item{port}{, hive or impala port}

\item{delim}{, optionnal argument for the delimiter of the file. Default to ',' .}

\item{hdfs_uri}{, url of platform}
}
\value{
nada , allows impala/hive table creation from 1 folder with csv(s) to Impala (creation, append)
}
\description{
df_2_impala
}
