% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ptools_utility_functions.r
\name{f2n}
\alias{f2n}
\title{Factor to numeric}
\usage{
f2n(fctr)
}
\arguments{
\item{fctr}{factor that we want to convert to numeric}
}
\value{
numeric vector
}
\description{
\code{f2n} returns a numeric vector, converted from factor
}
\details{
Returns a pure numeric vector
}
\examples{
set.seed(1234)
fctr <- factor(sample(1:4, 50, replace=TRUE), levels=1:4)
fctr
f2n(fctr)
}
\keyword{f2n}

