% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestPlot.R
\name{forestDMG}
\alias{forestDMG}
\title{Draw forest plot for differences betweeen cohorts.}
\usage{
forestDMG(
  m1,
  m2,
  m1Name = NULL,
  m2Name = NULL,
  minMut = 5,
  pVal = 0.05,
  fdr = NULL,
  color = NULL,
  geneFontSize = 1.2,
  titleSize = 1.2,
  lineWidth = 2.2
)
}
\arguments{
\item{m1}{first MAF object}

\item{m2}{second MAF object}

\item{m1Name}{optional name for first cohort}

\item{m2Name}{optional name for second cohort}

\item{minMut}{Consider only genes with minimum this number of samples mutated in atleast one of the cohort for analysis. Helful to ignore single mutated genes. Default 5.}

\item{pVal}{p-value threshold. Default 0.05.}

\item{fdr}{fdr threshold. Default NULL. If provided uses adjusted pvalues (fdr).}

\item{color}{vector of colors for cohorts. Default NULL.}

\item{geneFontSize}{Font size for gene symbols. Default 1.2}

\item{titleSize}{font size for titles. Default 1.2}

\item{lineWidth}{line width for CI bars. Default 2.2}
}
\description{
Draw forest plot for mutation differences analysis betweeen two cohorts.
}
\details{
The  x-axis is log10 converted odds ratio and y-axis is differentially mutated genes
}
\examples{
primary.apl <- system.file("extdata", "APL_primary.maf.gz", package = "maftools")
relapse.apl <- system.file("extdata", "APL_relapse.maf.gz", package = "maftools")
forestDMG(primary.apl, relapse.apl, m1Name = "Primary", m2Name = "Relapse", lineWidth = 3)
}
