% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptools_financialfunctions.r
\name{gaip}
\alias{gaip}
\title{Graduated annuity initial payment (payments grow at constant annual rate).}
\usage{
gaip(p, i, n, g, end)
}
\arguments{
\item{p}{principal}

\item{i}{interest rate}

\item{n}{number of periods}

\item{g}{growth rate of payments (annual)}

\item{end}{if TRUE, payment at the end of period}
}
\value{
The initial payment in the graduated annuity
}
\description{
\code{gaip} initial payment for a graduated annuity growing at rate g
}
\details{
Returns POSITIVE payment for positive principal
}
\examples{
gaip(100, 0.08, 10, 0.04)
gaip(100, 0.08, 10, 0.02, end = TRUE)
}
\keyword{gaip}
