% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptools_annuity_functions.r
\name{get_tla2}
\alias{get_tla2}
\title{Calculate temporary annuity values from a fixed entry age y to x (fixed start)}
\usage{
get_tla2(px, i, sx)
}
\arguments{
\item{px}{a vector of composite survivial probs from age x to x + n - 1. Length = n}

\item{i}{discount rate, scalar}

\item{sx}{salary scale. default is an n vector of 1, meaning no salary scale.}
}
\value{
an n vector storing the value of temporary life annuities from age x to age x + n - 1.
}
\description{
\code{get_tla2} present value of annuity-immediate
}
\details{
Suppose the age corresponding to px runs from a1 to aN, y = a1 (eg. age 30:65, y = 30)
This function conputes a..{y, x - y} and s_a..{y, x - y}, x ruuning from a1 to aN. 

Note that when x = a1 = y, we define a..{y, 0} = 0. so the first element is always 0. 
For age x > y, the number of years receiviing annuity is x - y, the resulting annuity value will be placed at age x. 
eg1: x = 31, y = 30,  annuity ($1) received only once at age 30, but the resulting annuity value will be placed at age 31.
eg2: x = 65, y = 30, annuity received from age 30 to 64(total 65 years), the resulting annuity value will be placed at age 65

Note that the last 2 survival rates and last salary scale are redundant in the calculation, they are just used as place holders. 
calculating the value of annuity running from 30 to 64 only involves survival rate from 30 to 63, 
because the last annuity payment is paid at the begining of 64. )
}
\examples{
get_tla2(rep(0.98, 65), 0.08, rep(1.1, 65))
}
\keyword{get_tla2}
