% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptools_geomean_functions.r
\name{gm.lr_to_r}
\alias{gm.lr_to_r}
\title{Find the arithmetic mean needed to result in a particular long-run geometric mean gm}
\usage{
gm.lr_to_r(gm, sd)
}
\arguments{
\item{gm}{Numeric vector of long-run geometric-mean returns.}

\item{sd}{standard deviation of the arithmetic mean}
}
\value{
The approximated arithmetic mean, given \code{gm} and \code{sd}.
}
\description{
Use brute force by minimizing error
}
\examples{
gm.lr_to_r(0.07331263, .12)
# compare to exact calculation
gm_to_r(0.07331263, .12)
}
