% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptools_geomean_functions.r
\name{gm_to_r}
\alias{gm_to_r}
\title{Get arithmetic mean associated with a long-run geometric mean}
\usage{
gm_to_r(gm, sd)
}
\arguments{
\item{gm}{Numeric vector of long-run geometric-mean returns.}

\item{sd}{standard deviation}
}
\value{
The arithmetic mean, given \code{gm} and \code{sd}.
}
\description{
This is the exact calculation rather than the ar = gr + var /2 approximation
See https://web.stanford.edu/~wfsharpe/mia/rr/mia_rr3.htm
}
\examples{
gm_to_r(0.07331263, .12)
}
