% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{hex_area}
\alias{hex_area}
\title{Get area of hexagon given length of side}
\usage{
hex_area(side)
}
\arguments{
\item{side}{scaler}
}
\value{
A scaler for the width
}
\description{
The length of the side is half of the length from vertex to vertex (so height in \code{geom_hex}).
}
\details{
For use with ggplot and \code{geom_hex} binwidth arguments, which expects arguments in width/height. I want hexagons in maps to be a specific area.
}
\examples{
area_check <- 1000
wh <- hex_dim(area_check^2)   #e.g. a square kilometer if spatial units are in meters
area <- hex_area(wh[1]/2)       #inverse operation
all.equal(area_check,sqrt(area))
wi <- hex_wd(wh[1])
all.equal(wh[2],wi)

}
\seealso{
\code{\link[=hex_wd]{hex_wd()}} for estimating the width given the height
\code{\link[=hex_dim]{hex_dim()}} for estimating width/height given area
}
