% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ptools_utility_functions.r
\name{ht}
\alias{ht}
\title{Show head and tail of a vector, matrix, table, data frame or function}
\usage{
ht(df, nrecs=6)
}
\arguments{
\item{df}{The object. No default.}

\item{nrecs}{number of records, rows, whatever to show at head and at tail}
}
\description{
\code{ht} head and tail of a vector, matrix, table, data frame or function
}
\details{
show head and tail of a vector, matrix, table, data frame or function
}
\examples{
ht(mtcars, 4)
}
\keyword{ht}

