% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{list_files}
\alias{list_files}
\title{list_files}
\usage{
list_files(hdfsUri, dirUri)
}
\arguments{
\item{hdfsUri, }{the Uniform Resource Identifier of the HDFS used}

\item{dirUri, }{the path to the directory to list files from}
}
\value{
a vector with the file names from corresponding directory.
}
\description{
list_files
}
