% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ptools_utility_functions.r
\name{memory}
\alias{memory}
\title{Describe memory usage and collect garbage}
\usage{
memory(maxnobjs=5)
}
\arguments{
\item{maxnobjs}{The number of objects to display. Default is 5.}
}
\description{
\code{memory} describe memory usage and collect garbage
}
\details{
Describes memory usage and collects garbage
}
\examples{
memory(4)
}
\keyword{memory}

