% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_files.R
\name{move_files}
\alias{move_files}
\title{move_files}
\usage{
move_files(hdfsUri, dirUri, destUri, user_name)
}
\arguments{
\item{hdfsUri, }{the Uniform Resource Identifier of the HDFS used}

\item{dirUri, }{the path to the directory to move files from}

\item{destUri, }{the path to the directory to move files to}

\item{user_name, }{a user name with write and delelete rights}
}
\value{
Nothing. All files within directory will be moved.
}
\description{
move_files
}
