% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.varImpPlot.R
\name{my.varImpPlot}
\alias{my.varImpPlot}
\title{my variable importance plot}
\usage{
my.varImpPlot(
  imp,
  sort = TRUE,
  n.var = min(30, length(imp)),
  type = NULL,
  class = NULL,
  scale = TRUE,
  main = " ",
  ...
)
}
\arguments{
\item{imp}{importance values}
}
\value{
S
}
\description{
Random Forest variable importance plot
}
