% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_bor}
\alias{nyc_bor}
\title{NYC Boroughs}
\format{
A SpatialPolygonsDataFrame object of the NYC Boroughs. This is projected (same coordinates as shootings). See the \href{https://www.nyc.gov/site/planning/data-maps/open-data.page}{Bytes of the Big Apple} for any details on the file.
}
\source{
\itemize{
\item \url{https://www1.nyc.gov/assets/planning/download/zip/data-maps/open-data/nybb_21c.zip}
}
}
\usage{
nyc_bor
}
\description{
Spatial file for New York City Borough outlines without water areas
}
\keyword{datasets}
