% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_cafe}
\alias{nyc_cafe}
\title{NYC Sidewalk Cafes}
\format{
A SpatialPointsDataFrame with point locations Sidwalk cafes in NYC. Note currently
includes only active license locations. Current N around 400 and none in Staten Island.
}
\source{
\itemize{
\item \url{https://data.cityofnewyork.us/api/views/qcdj-rwhu/rows.csv?accessType=DOWNLOAD}
}
}
\usage{
nyc_cafe
}
\description{
Point locations for sidewalk cafes in NYC
}
\keyword{datasets}
