% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_shoot}
\alias{nyc_shoot}
\title{NYPD Open Data on Shootings}
\format{
A data frame with currently over 20k rows and 21 fields, including date/time and address level geocoordinates for the event. Data from 2006 to currently. See the info on Socrata for the field name codebook.
}
\source{
\itemize{
\item \url{https://data.cityofnewyork.us/Public-Safety/NYPD-Shooting-Incident-Data-Year-To-Date-/5ucz-vwe8} for current
\item \url{https://data.cityofnewyork.us/Public-Safety/NYPD-Shooting-Incident-Data-Historic-/833y-fsy8} for historical
}
}
\usage{
nyc_shoot
}
\description{
Shootings recorded from the New York City Police Department from 2006 to current.
}
\keyword{datasets}
