% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ptools_financialfunctions.r
\name{pdiff}
\alias{pdiff}
\title{Percent difference, first value relative to second}
\usage{
pdiff(first, second)
}
\arguments{
\item{first}{first value}

\item{second}{second value}
}
\value{
The percent difference
}
\description{
\code{pdiff} % difference, first value vs. second value
}
\details{
Use, for example, to get percent difference of actual assets from expected assets
}
\examples{
pdiff(100, 90)
}
\keyword{pdiff}

