% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pdiffb}
\alias{pdiffb}
\title{Percent difference, first value relative to second}
\usage{
pdiffb(first, second)
}
\arguments{
\item{first}{first value}

\item{second}{second value}
}
\value{
The percent difference
}
\description{
\code{pdiffb} % difference, first value vs. second value
}
\details{
Use, for example, to get percent difference of actual assets from expected assets
}
\examples{
pdiffb(100, 90)
}
\keyword{pdiffb}

