% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ptools_financialfunctions.r
\name{pmt}
\alias{pmt}
\title{Amortization function}
\usage{
pmt(p, i, n)
}
\arguments{
\item{p}{principal}

\item{i}{interest rate}

\item{n}{number of periods}
}
\value{
The level payment in the annuity
}
\description{
\code{pmt} level amortization for principal p, interest rate i, n periods
}
\details{
Returns POSITIVE payment for positive principal
}
\examples{
pmt(450, .07, 20)
}
\keyword{pmt}

