% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptools_financialfunctions.r
\name{pvann}
\alias{pvann}
\title{Present value of an annuity-immediate (with pmt at end of period)}
\usage{
pvann(i, n, pmt)
}
\arguments{
\item{i}{interest rate}

\item{n}{number of periods}

\item{pmt}{periodic payment}
}
\value{
The initial payment
}
\description{
\code{pvann} present value of annuity-immediate
}
\details{
Returns a POSITIVE present value for positive payment
}
\examples{
pvann(.05, 30, 100)
}
\keyword{pvann}
