% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptools_geomean_functions.r
\name{r_to_gm.n}
\alias{r_to_gm.n}
\title{Get expected geometric mean at year n, for a given arithmentic mean r}
\usage{
r_to_gm.n(r, sd, n)
}
\arguments{
\item{r}{Numeric vector of arithmetic returns}

\item{sd}{standard deviation}

\item{n}{year}
}
\value{
Expected geometric mean at year \code{n}, given \code{r} and \code{sd}.
}
\description{
Get expected geometric mean at year n, for a given arithmentic mean r
}
\examples{
r_to_gm.n(.08, .12, 1)
r_to_gm.n(.08, .12, 10)
r_to_gm.n(.08, .12, 100)
r_to_gm.n(.08, .12, 1000000)
# compare to the exact calculation -- hmmm...
r_to_gm(.08, .12)
# compare to the approximation
.08 - .12^2 / 2
}
