% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptools_geomean_functions.r
\name{r_to_gm}
\alias{r_to_gm}
\title{Get long-run geometric mean associated with arithmetic mean}
\usage{
r_to_gm(r, sd)
}
\arguments{
\item{r}{Numeric vector of arithmetic returns.}

\item{sd}{standard deviation}
}
\value{
The long-run geometric mean, given \code{r} and \code{sd}.
}
\description{
This is the exact calculation rather than the gm = r - var /2 approximation.
See https://web.stanford.edu/~wfsharpe/mia/rr/mia_rr3.htm
}
\examples{
r_to_gm(.08, .12)
}
