% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_RFadaptor.R
\name{ranger_RFadaptor}
\alias{ranger_RFadaptor}
\title{function to translate ranger to pseudo randomForest}
\usage{
ranger_RFadaptor(ra, y = NULL)
}
\arguments{
\item{ra}{ranger object}
}
\value{
S
}
\description{
Random Forest variable importance plot
}
