% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankedDotsByGroups.R
\name{rankedDotsByGroups}
\alias{rankedDotsByGroups}
\title{Draw dot plot with ranked order}
\usage{
rankedDotsByGroups(x, col = NULL, alpha = 1, args.legend = NULL, ...)
}
\arguments{
\item{x}{a numeric vector or a list of numeric vectors}

\item{col}{a vectors of color codes for each groups}

\item{alpha}{the opacity of color}

\item{args.legend}{list of additional arguments to pass to \code{legend()}.}

\item{...}{arguments to be passed to \code{plot()}.}
}
\description{
This function draws dot plot with ranked order, if multiple
groups exists, dots is ranked within each groups.
}
\examples{
set.seed(123)
z <- list(
  NAT = round(rnorm(200, mean = 7, sd = 1) * 1000),
  Tumor = round(rnorm(200, mean = 9, sd = 1) * 1000)
)
rankedDotsByGroups(z, xlab = "Sample Number", pch=19, main = "Number of identified proteins")
}
