\name{Undocumented}

\alias{AFFYLookup}
\alias{AFFYMedianNorm}
\alias{AFFYQuantileNorm}
\alias{PairsPlot}
\alias{expval}
\alias{genegroup}
\alias{permute}
\alias{plotdiag}
\alias{post.fastRP}

\title{ Undocumented objects }

\description{
All the objects on this page do not have proper documentation, but in order
to pass \code{R CMD check} I have listed them here.
}

\author{ Tim Keighley }
\keyword{ internal }
