% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_publication.R
\name{scale_fill_publication}
\alias{scale_fill_publication}
\title{scale fill colors}
\usage{
scale_fill_publication(colors = NULL, ...)
}
\arguments{
\item{colors}{a vector of color codes}

\item{...}{other parameters pass to discrete_scale() function}
}
\description{
scale fill colors (barplot,boxplot...) with manually specified colors
}
\seealso{
\code{\link[=scale_colour_publication]{scale_colour_publication()}} \code{\link[=scale_fill_publication_dark]{scale_fill_publication_dark()}} \code{\link[=scale_colour_publication_dark]{scale_colour_publication_dark()}}
}
