% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_publication.R
\name{scale_fill_publication_dark}
\alias{scale_fill_publication_dark}
\title{Title}
\usage{
scale_fill_publication_dark(colors = NULL, ...)
}
\arguments{
\item{colors}{a vector of color codes}

\item{...}{other parameters pass to discrete_scale() function}
}
\description{
Title
}
