% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_publication.R
\name{theme_publication}
\alias{theme_publication}
\title{A ggplot2 complete theme for make publication ready figures}
\usage{
theme_publication(base_size = 14, base_family = "sans")
}
\arguments{
\item{base_size}{font size, default 14}

\item{base_family}{font family, defaut 'sans'}
}
\description{
A ggplot2 complete theme for make publication ready figures
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, disp)) + geom_point() + theme_publication()
}
\author{
Koundinya Desiraju (https://github.com/koundy/ggplot_theme_Publication)
}
