% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_files.R
\name{copy_files}
\alias{copy_files}
\title{copy_files}
\usage{
copy_files(hdfsUri, dirUri, destUri, user_name, copy_all = TRUE, fname = "",
  from_hdfs = TRUE, from_folder = "", newname = "")
}
\arguments{
\item{hdfsUri}{, the url of the platform}

\item{dirUri}{, the path to the folder of origin}

\item{destUri}{, the path to the folder of destination}

\item{user_name}{, credential with writing credential}

\item{copy_all}{, optionnal argument : shall all the files in origin folder be copied ?}

\item{fname}{, optionnal argument : name of the unique file to copy}

\item{from_hdfs}{, optionnal argument : is the file stored locally in R session ?}

\item{from_folder}{, optionnal argument : boolean, is the file to move stored in a local folder ?}

\item{newname}{, optionnal default to "", usefull when copy from HDFS : allow copied file renaming}
}
\value{
nada , the file is copied to wanted directory
}
\description{
copy_files
}
