% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_files.R
\name{delete_files}
\alias{delete_files}
\title{delete_files}
\usage{
delete_files(hdfsUri, dirUri, user_name)
}
\arguments{
\item{hdfsUri, }{the Uniform Resource Identifier of the HDFS used}

\item{dirUri, }{the path to the directory to delete files from}

\item{user_name, }{a user name with write and delelete rights}
}
\value{
Nothing. Files in directory are removed.
}
\description{
delete_files
}
