% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ptools_financialfunctions.r
\name{gaip}
\alias{gaip}
\title{Graduated annuity initial payment (payments grow at constant annual rate).}
\usage{
gaip(p, i, g, n)
}
\arguments{
\item{p}{principal}

\item{i}{interest rate}

\item{g}{growth rate of payments (annual)}

\item{n}{number of periods}
}
\value{
The initial payment in the graduated annuity
}
\description{
\code{gaip} initial payment for a graduated annuity growing at rate g
}
\details{
Returns POSITIVE payment for positive principal
}
\examples{
gaip(450, .07, .03, 20)
}
\keyword{gaip}

