% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{list_files}
\alias{list_files}
\title{List files from HDFS folder.}
\usage{
list_files(hdfsUri, dirUri)
}
\arguments{
\item{hdfsUri}{, hdfs url}

\item{dirUri}{, path to folder}
}
\value{
a character vector of files in directory
}
\description{
List files from HDFS folder.
}
