% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_files.R
\name{move_files}
\alias{move_files}
\title{Move files from HDFS folders.}
\usage{
move_files(hdfsUri, dirUri, destUri, user_name, move_all = TRUE,
  fname = "", newname = "")
}
\arguments{
\item{hdfsUri, }{hdfs url}

\item{dirUri, }{path to directory to move file from}

\item{destUri, }{destination folder}

\item{user_name, }{credential with write access}

\item{move_all, }{boolean : should all file in folder been moved ?}

\item{fname, }{name of the file to move}

\item{newname, }{default to empty string (no change). Allows to rename the file if needed.}
}
\value{
nothing, see more HDFS API at ttps://hadoop.apache.org/docs/r1.0.4/webhdfs.html
}
\description{
Move files from HDFS folders.
}
