% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_liq}
\alias{nyc_liq}
\title{NYC Alcohol Licenses}
\format{
A SpatialPointsDataFrame with point locations for alcohol licenses inside of NYC.
Note that some of these are not the actual sales place, but another address for the business.
Currently over 18,000 addresses.
}
\source{
\itemize{
\item \url{https://data.ny.gov/Economic-Development/Liquor-Authority-Current-List-of-Active-Licenses/hrvs-fxs2}
}
}
\usage{
nyc_liq
}
\description{
Point locations for alcohol locations inside NYC boroughs
}
\keyword{datasets}
